// app/src/Pages/__generated__/findManyRole.ts
/* tslint:disable */
/* eslint-disable */
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: findManyRole
// ====================================================

export interface findManyRole_findManyRole {
  __typename: "Role";
  id: string;
  slug: string;
}

export interface findManyRole {
  findManyRole: findManyRole_findManyRole[] | null;
}